reworked scaling

+++ SelectedTextTranslate/Source/View/Controls/Buttons/HoverFlatButtonWindow.h
@@ -40,5 +40,5 @@ public:
 
     Size GetComputedSize() const;
 
-    double GetTextBaseline() const;
+    int GetTextBaseline() const;
 };
